\name{RcmdrSCDAPlugin-internal}
\title{Internal RcmdrSCDAPlugin objects}
\alias{Rcmdr.quantity}
\alias{Rcmdr.selectdesign}
\alias{Rcmdr.assignments}
\alias{Rcmdr.graph}
\alias{Rcmdr.observed}
\alias{Rcmdr.distribution}
\alias{Rcmdr.pvalue}
\alias{Rcmdr.ES}
\alias{Rcmdr.combine}
\alias{Rcmdr.graph.CL}
\alias{Rcmdr.graph.VAR}
\alias{Rcmdr.graph.TREND}
\description{Internal RcmdrSCDAPlugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}

