\name{RcmdrSCDAPlugin-package}
\alias{RcmdrSCDAPlugin-package}
\alias{RcmdrSCDAPlugin}
\docType{package}
\title{
Install the SCDA Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" for the SCVA, SCRT and SCMA packages. 
}
\details{
\tabular{ll}{
Package: \tab RcmdrSCDAPlugin\cr
Type: \tab Package\cr
Version: \tab O.0-O\cr
Date: \tab 2012-04-06\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Isis Bulte and Patrick Onghena

Maintainer: Isis Bulte <isis.bulte@ppw.kuleuven.be>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

