\name{RcmdrSCDAPlugin-package}
\alias{RcmdrSCDAPlugin-package}
\alias{RcmdrSCDAPlugin}
\docType{package}
\title{
Install the SCDA Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" for the SCVA, SCRT and SCMA packages. 
}
\details{
\tabular{ll}{
Package: \tab RcmdrSCDAPlugin\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-13\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Isis Bulte and Patrick Onghena

Maintainer: Marlies Vervloet <marlies.vervloet@ppw.kuleuven.be>
}
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}
\keyword{ package }
